Shift SRD monster functions to separate module
